requirements.txt
pyqrcode==1.2.1
onetimepass==1.0.1
flask-wtf==0.14.3
flask-bootstrap==3.3.7.1
flask-sqlalchemy==2.4.4
flask-login==0.5.0
python-dateutil==2.8.1
boto3==1.14.33
colorlog==4.2.1